refactor: remove unused onAssetsOpen prop and related code from input components and dialogs
