+++ ui/src/components/SchemaForm/Templates/BaseInputTemplate/ColorInput.tsx
@@ -17,7 +17,6 @@ const ColorInput = <
 >(
   props: BaseInputTemplateProps<T, S, F> & {
     onEditorOpen?: () => void;
-    onAssetsOpen?: () => void;
   },
 ) => {
   const {
@@ -32,7 +31,6 @@ const ColorInput = <
     onBlur,
     onFocus,
     onEditorOpen,
-    onAssetsOpen,
     options,
     schema,
     rawErrors = [],
@@ -89,7 +87,6 @@ const ColorInput = <
           value={value}
           defaultValue={defaultValue}
           onEditorOpen={onEditorOpen}
-          onAssetsOpen={onAssetsOpen}
           onReset={handleReset}
         />
       </div>
