+++ ui/src/components/SchemaForm/index.tsx
@@ -21,7 +21,6 @@ type SchemaFormProps = {
   onError?: (errors: RJSFValidationError[]) => void;
   onValidationChange?: (isValid: boolean) => void;
   onEditorOpen?: (fieldContext: FieldContext) => void;
-  onAssetsOpen?: (fieldContext: FieldContext) => void;
 };
 
 const SchemaForm: React.FC<SchemaFormProps> = ({
@@ -32,7 +31,6 @@ const SchemaForm: React.FC<SchemaFormProps> = ({
   onError,
   onValidationChange,
   onEditorOpen,
-  onAssetsOpen,
 }) => {
   const t = useT();
   const [error, setError] = useState<string | null>(null);
@@ -91,7 +89,7 @@ const SchemaForm: React.FC<SchemaFormProps> = ({
         formData={defaultFormData}
         validator={validator}
         uiSchema={{ "ui:submitButtonOptions": { norender: true } }} // We handle submissions outside of this component
-        formContext={{ onEditorOpen, onAssetsOpen }}
+        formContext={{ onEditorOpen }}
         onChange={handleChange}
         onError={handleError}
       />
