+++ ui/src/features/Editor/components/ParamsDialog/components/ParamEditor/index.tsx
@@ -36,7 +36,6 @@ type Props = {
   ) => Promise<void>;
   onWorkflowRename?: (id: string, name: string) => void;
   onValueEditorOpen: (fieldContext: FieldContext) => void;
-  onAssetsOpen?: (fieldContext: FieldContext) => void;
 };
 
 const ParamEditor: React.FC<Props> = ({
@@ -49,7 +48,6 @@ const ParamEditor: React.FC<Props> = ({
   onUpdate,
   onWorkflowRename,
   onValueEditorOpen,
-  onAssetsOpen,
 }) => {
   const t = useT();
   const { useGetActionById } = useAction(i18n.language);
@@ -158,7 +156,6 @@ const ParamEditor: React.FC<Props> = ({
                   onChange={onParamsUpdate}
                   onValidationChange={handleParamsValidationChange}
                   onEditorOpen={onValueEditorOpen}
-                  onAssetsOpen={onAssetsOpen}
                 />
               )}
             </div>
