+++ ui/src/features/Editor/components/ParamsDialog/index.tsx
@@ -8,9 +8,8 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@flow/components";
-import AssetsDialog from "@flow/features/AssetsDialog";
 import { useT } from "@flow/lib/i18n";
-import { Asset, Node } from "@flow/types";
+import type { Node } from "@flow/types";
 
 import { ParamEditor, ValueEditorDialog } from "./components";
 import { FieldContext, setValueAtPath } from "./utils/fieldUtils";
@@ -37,7 +36,6 @@ const ParamsDialog: React.FC<Props> = ({
   const t = useT();
 
   const [openValueEditor, setOpenValueEditor] = useState(false);
-  const [showAssets, setShowAssets] = useState(false);
   const [currentFieldContext, setCurrentFieldContext] = useState<
     FieldContext | undefined
   >(undefined);
@@ -98,11 +96,6 @@ const ParamsDialog: React.FC<Props> = ({
     }
   };
 
-  const handleAssetDoubleClick = (asset: Asset) => {
-    const v = asset.url;
-    handleValueChange(v);
-  };
-
   return (
     <>
       <Dialog open={!!openNode} onOpenChange={() => onOpenNode()}>
@@ -129,10 +122,6 @@ const ParamsDialog: React.FC<Props> = ({
                 setCurrentFieldContext(fieldContext);
                 setOpenValueEditor(true);
               }}
-              onAssetsOpen={(fieldContext) => {
-                setCurrentFieldContext(fieldContext);
-                setShowAssets(true);
-              }}
             />
           )}
         </DialogContent>
@@ -146,15 +135,6 @@ const ParamsDialog: React.FC<Props> = ({
             setCurrentFieldContext(undefined);
           }}
           onValueSubmit={handleValueChange}
-        />
-      )}
-      {showAssets && currentFieldContext && (
-        <AssetsDialog
-          onDialogClose={() => {
-            setShowAssets(false);
-            setCurrentFieldContext(undefined);
-          }}
-          onAssetDoubleClick={handleAssetDoubleClick}
         />
       )}
     </>
